+++ oneflow/user/kernels/upsample_bicubic_2d_kernel.cu
@@ -145,8 +145,8 @@ class UpsampleBicubic2dGPUKernel final : public user_op::OpKernel {
     const int64_t in_width = x_tensor->shape().At(3);
     const int64_t out_height = y_tensor->shape().At(2);
     const int64_t out_width = y_tensor->shape().At(3);
-    const float height_scale = out_height * 1.0 / in_height;
-    const float width_scale = out_width * 1.0 / in_width;
+    const double height_scale = out_height * 1.0 / in_height;
+    const double width_scale = out_width * 1.0 / in_width;
     const int64_t elem_cnt = out_height * out_width;
 
     if (in_height == out_height && in_width == out_width) {
@@ -186,8 +186,8 @@ class UpsampleBicubic2dGradGPUKernel final : public user_op::OpKernel {
     const int64_t in_width = dx_tensor->shape().At(3);
     const int64_t out_height = dy_tensor->shape().At(2);
     const int64_t out_width = dy_tensor->shape().At(3);
-    const float height_scale = out_height * 1.0 / in_height;
-    const float width_scale = out_width * 1.0 / in_width;
+    const double height_scale = out_height * 1.0 / in_height;
+    const double width_scale = out_width * 1.0 / in_width;
     const int64_t elem_cnt = out_height * out_width;
 
     if (in_height == out_height && in_width == out_width) {
