+++ oneflow/user/kernels/upsample_bilinear_2d_kernel.cpp
@@ -98,8 +98,8 @@ class UpsampleBilinear2DCPUKernel final : public user_op::OpKernel {
     const int64_t in_width = x_tensor->shape().At(3);
     const int64_t out_height = y_tensor->shape().At(2);
     const int64_t out_width = y_tensor->shape().At(3);
-    const float height_scale = out_height * 1.0 / in_height;
-    const float width_scale = out_width * 1.0 / in_width;
+    const double height_scale = out_height * 1.0 / in_height;
+    const double width_scale = out_width * 1.0 / in_width;
 
     if (in_height == out_height && in_width == out_width) {
       memcpy(y_tensor->mut_dptr<void>(), x_tensor->dptr<void>(),
@@ -140,9 +140,9 @@ class UpsampleBilinear2DGradCPUKernel final : public user_op::OpKernel {
     const int64_t in_width = dx_tensor->shape().At(3);
     const int64_t out_height = dy_tensor->shape().At(2);
     const int64_t out_width = dy_tensor->shape().At(3);
-    const float height_scale = out_height * 1.0 / in_height;
-    const float width_scale = out_width * 1.0 / in_width;
-    
+    const double height_scale = out_height * 1.0 / in_height;
+    const double width_scale = out_width * 1.0 / in_width;
+
     if (in_height == out_height && in_width == out_width) {
       memcpy(dx_tensor->mut_dptr<void>(), dy_tensor->dptr<void>(),
              sizeof(T) * nbatch * channels * in_height * in_width);
